requirements.txt
numpy
pillow
scikit-image
torch
torchvision
